chore: Break circular dependency between themes' color palettes
