+++ style-dictionary/visual-refresh/colors.ts
@@ -1,6 +1,6 @@
 // Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
 // SPDX-License-Identifier: Apache-2.0
-import { paletteTokens } from '../core/color-palette.js';
+import { paletteTokens } from '../core/palette-values.js';
 import { expandColorDictionary } from '../utils/index.js';
 import { StyleDictionary } from '../utils/interfaces.js';
